+++ src/Curry/Syntax/Parser.hs
@@ -13,14 +13,10 @@
     The Curry parser is implemented using the (mostly) LL(1) parsing
     combinators implemented in 'Curry.Base.LLParseComb'.
 -}
-{-# LANGUAGE CPP #-}
 module Curry.Syntax.Parser
   ( parseSource, parseHeader, parseInterface, parseGoal
   ) where
 
-#if __GLASGOW_HASKELL__ >= 710
-import Prelude hiding ((<$>), (<*>))
-#endif
 import Curry.Base.Ident
 import Curry.Base.Monad       (CYM)
 import Curry.Base.Position    (Position, mk, mk')
@@ -56,7 +52,7 @@ parseGoal = fullParser goal lexer ""
 -- ---------------------------------------------------------------------------
 
 -- |Parser for a module header
-moduleHeader :: Parser Token ([ImportDecl] -> [Decl] -> Module) a
+moduleHeader :: Parser a Token ([ImportDecl] -> [Decl] -> Module)
 moduleHeader = (\ps (m, es) -> Module ps m es)
            <$> modulePragmas
            <*> header
@@ -65,17 +61,17 @@ moduleHeader = (\ps (m, es) -> Module ps m es)
                      <*-> expectWhere
                 `opt` (mainMIdent, Nothing)
 
-modulePragmas :: Parser Token [ModulePragma] a
+modulePragmas :: Parser a Token [ModulePragma]
 modulePragmas = many (languagePragma <|> optionsPragma)
 
-languagePragma :: Parser Token ModulePragma a
+languagePragma :: Parser a Token ModulePragma
 languagePragma =   LanguagePragma
               <$>  tokenPos PragmaLanguage
               <*>  (languageExtension `sepBy1` comma)
               <*-> token PragmaEnd
   where languageExtension = classifyExtension <$> ident
 
-optionsPragma :: Parser Token ModulePragma a
+optionsPragma :: Parser a Token ModulePragma
 optionsPragma = (\pos a -> OptionsPragma pos (fmap classifyTool $ toolVal a)
                                              (toolArgs a))
            <$>  position
@@ -83,25 +79,25 @@ optionsPragma = (\pos a -> OptionsPragma pos (fmap classifyTool $ toolVal a)
            <*-> token PragmaEnd
 
 -- |Parser for an export specification
-exportSpec :: Parser Token ExportSpec a
+exportSpec :: Parser a Token ExportSpec
 exportSpec = Exporting <$> position <*> parens (export `sepBy` comma)
 
 -- |Parser for an export item
-export :: Parser Token Export a
+export :: Parser a Token Export
 export =  qtycon <**> (parens spec `opt` Export)         -- type constructor
       <|> Export <$> qfun <\> qtycon                     -- fun
       <|> ExportModule <$-> token KW_module <*> modIdent -- module
   where spec =       ExportTypeAll  <$-> token DotDot
             <|> flip ExportTypeWith <$>  con `sepBy` comma
 
-moduleDecls :: Parser Token ([ImportDecl], [Decl]) a
+moduleDecls :: Parser a Token ([ImportDecl], [Decl])
 moduleDecls = impDecl <$> importDecl
                       <*> (semicolon <-*> moduleDecls `opt` ([], []))
           <|> (,) []  <$> topDecls
   where impDecl i (is, ds) = (i:is ,ds)
 
 -- |Parser for a single import declaration
-importDecl :: Parser Token ImportDecl a
+importDecl :: Parser a Token ImportDecl
 importDecl =  flip . ImportDecl
           <$> tokenPos KW_import
           <*> flag (token Id_qualified)
@@ -110,7 +106,7 @@ importDecl =  flip . ImportDecl
           <*> option importSpec
 
 -- |Parser for an import specification
-importSpec :: Parser Token ImportSpec a
+importSpec :: Parser a Token ImportSpec
 importSpec =   position
           <**> (Hiding <$-> token Id_hiding `opt` Importing)
           <*>  parens (spec `sepBy` comma)
@@ -125,33 +121,33 @@ importSpec =   position
 -- ---------------------------------------------------------------------------
 
 -- |Parser for an interface
-interface :: Parser Token Interface a
+interface :: Parser a Token Interface
 interface = uncurry <$> intfHeader <*> braces intfDecls
 
-intfHeader :: Parser Token ([IImportDecl] -> [IDecl] -> Interface) a
+intfHeader :: Parser a Token ([IImportDecl] -> [IDecl] -> Interface)
 intfHeader = Interface <$-> token Id_interface <*> modIdent <*-> expectWhere
 
-intfDecls :: Parser Token ([IImportDecl], [IDecl]) a
+intfDecls :: Parser a Token ([IImportDecl], [IDecl])
 intfDecls = impDecl <$> iImportDecl
                     <*> (semicolon <-*> intfDecls `opt` ([], []))
         <|> (,) [] <$> intfDecl `sepBy` semicolon
   where impDecl i (is, ds) = (i:is, ds)
 
 -- |Parser for a single interface import declaration
-iImportDecl :: Parser Token IImportDecl a
+iImportDecl :: Parser a Token IImportDecl
 iImportDecl = IImportDecl <$> tokenPos KW_import <*> modIdent
 
 -- |Parser for a single interface declaration
-intfDecl :: Parser Token IDecl a
+intfDecl :: Parser a Token IDecl
 intfDecl = choice [ iInfixDecl, iHidingDecl, iDataDecl, iNewtypeDecl
                   , iTypeDecl , iFunctionDecl <\> token Id_hiding ]
 
 -- |Parser for an interface infix declaration
-iInfixDecl :: Parser Token IDecl a
+iInfixDecl :: Parser a Token IDecl
 iInfixDecl = infixDeclLhs IInfixDecl <*> integer <*> qfunop
 
 -- |Parser for an interface hiding declaration
-iHidingDecl :: Parser Token IDecl a
+iHidingDecl :: Parser a Token IDecl
 iHidingDecl = tokenPos Id_hiding <**> (hDataDecl <|> hFuncDecl)
   where
   hDataDecl = hiddenData <$-> token KW_data <*> qtycon <*> many tyvar
@@ -161,22 +157,22 @@ iHidingDecl = tokenPos Id_hiding <**> (hDataDecl <|> hFuncDecl)
   hidingFunc a ty p = IFunctionDecl p (qualify (mkIdent "hiding")) a ty
 
 -- |Parser for an interface data declaration
-iDataDecl :: Parser Token IDecl a
+iDataDecl :: Parser a Token IDecl
 iDataDecl = iTypeDeclLhs IDataDecl KW_data <*> constrs <*> iHidden
   where constrs = equals <-*> constrDecl `sepBy1` bar `opt` []
 
 -- |Parser for an interface newtype declaration
-iNewtypeDecl :: Parser Token IDecl a
+iNewtypeDecl :: Parser a Token IDecl
 iNewtypeDecl = iTypeDeclLhs INewtypeDecl KW_newtype
                <*-> equals <*> newConstrDecl <*> iHidden
 
 -- |Parser for an interface type synonym declaration
-iTypeDecl :: Parser Token IDecl a
+iTypeDecl :: Parser a Token IDecl
 iTypeDecl = iTypeDeclLhs ITypeDecl KW_type
             <*-> equals <*> type0
 
 -- |Parser for an interface hiding pragma
-iHidden :: Parser Token [Ident] a
+iHidden :: Parser a Token [Ident]
 iHidden = token PragmaHiding
           <-*> (con `sepBy` comma)
           <*-> token PragmaEnd
@@ -184,44 +180,44 @@ iHidden = token PragmaHiding
 
 
 -- |Parser for an interface function declaration
-iFunctionDecl :: Parser Token IDecl a
+iFunctionDecl :: Parser a Token IDecl
 iFunctionDecl =  IFunctionDecl <$> position <*> qfun <*> arity
             <*-> token DoubleColon <*> type0
 
 -- |Parser for function's arity
-arity :: Parser Token Int a
+arity :: Parser a Token Int
 arity = int `opt` 0
 
 iTypeDeclLhs :: (Position -> QualIdent -> [Ident] -> a) -> Category
-             -> Parser Token a b
+             -> Parser b Token a
 iTypeDeclLhs f kw = f <$> tokenPos kw <*> qtycon <*> many tyvar
 
 -- ---------------------------------------------------------------------------
 -- Top-Level Declarations
 -- ---------------------------------------------------------------------------
 
-topDecls :: Parser Token [Decl] a
+topDecls :: Parser a Token [Decl]
 topDecls = topDecl `sepBy` semicolon
 
-topDecl :: Parser Token Decl a
+topDecl :: Parser a Token Decl
 topDecl = choice [ dataDecl, newtypeDecl, typeDecl
                  , foreignDecl, infixDecl, functionDecl ]
 
-dataDecl :: Parser Token Decl a
+dataDecl :: Parser a Token Decl
 dataDecl = typeDeclLhs DataDecl KW_data <*> constrs
   where constrs = equals <-*> constrDecl `sepBy1` bar `opt` []
 
-newtypeDecl :: Parser Token Decl a
+newtypeDecl :: Parser a Token Decl
 newtypeDecl = typeDeclLhs NewtypeDecl KW_newtype <*-> equals <*> newConstrDecl
 
-typeDecl :: Parser Token Decl a
+typeDecl :: Parser a Token Decl
 typeDecl = typeDeclLhs TypeDecl KW_type <*-> equals <*> type0
 
 typeDeclLhs :: (Position -> Ident -> [Ident] -> a) -> Category
-            -> Parser Token a b
+            -> Parser b Token a
 typeDeclLhs f kw = f <$> tokenPos kw <*> tycon <*> many anonOrTyvar
 
-constrDecl :: Parser Token ConstrDecl a
+constrDecl :: Parser a Token ConstrDecl
 constrDecl = position <**> (existVars <**> constr)
   where
   constr =  conId     <**> identDecl
@@ -241,23 +237,23 @@ constrDecl = position <**> (existVars <**> constr)
   conOpDeclPrefix op ty1 ty2 tvs p = ConOpDecl p tvs ty1 op ty2
   recDecl fs c tvs p               = RecordDecl p tvs c fs
 
-fieldDecl :: Parser Token FieldDecl a
+fieldDecl :: Parser a Token FieldDecl
 fieldDecl = FieldDecl <$> position <*> labels <*-> token DoubleColon <*> type0
   where labels = fun `sepBy1` comma
 
-newConstrDecl :: Parser Token NewConstrDecl a
+newConstrDecl :: Parser a Token NewConstrDecl
 newConstrDecl = position <**> (existVars <**> (con <**> newConstr))
   where newConstr =  newConDecl <$> type2
                  <|> newRecDecl <$> newFieldDecl
         newConDecl ty  c vs p = NewConstrDecl p vs c ty
         newRecDecl fld c vs p = NewRecordDecl p vs c fld
 
-newFieldDecl :: Parser Token (Ident, TypeExpr) a
+newFieldDecl :: Parser a Token (Ident, TypeExpr)
 newFieldDecl = layoutOff <-*> braces labelDecl
   where labelDecl = (,) <$> fun <*-> token DoubleColon <*> type0
 
 -- Parsing of existential variables (currently disabled)
-existVars :: Parser Token [Ident] a
+existVars :: Parser a Token [Ident]
 {-
 existVars flat
   | flat = succeed []
@@ -265,14 +261,14 @@ existVars flat
 -}
 existVars = succeed []
 
-functionDecl :: Parser Token Decl a
+functionDecl :: Parser a Token Decl
 functionDecl = position <**> decl
   where
   decl = fun `sepBy1` comma <**> funListDecl
     <|?> mkFunDecl <$> lhs <*> declRhs
   lhs = (\f -> (f, FunLhs f [])) <$> fun <|?> funLhs
 
-funListDecl :: Parser Token ([Ident] -> Position -> Decl) a
+funListDecl :: Parser a Token ([Ident] -> Position -> Decl)
 funListDecl =  typeSig           <$-> token DoubleColon <*> type0
            <|> flip ExternalDecl <$-> token KW_external
   where typeSig ty vs p = TypeSig p vs ty
@@ -280,7 +276,7 @@ funListDecl =  typeSig           <$-> token DoubleColon <*> type0
 mkFunDecl :: (Ident, Lhs) -> Rhs -> Position -> Decl
 mkFunDecl (f, lhs) rhs' p = FunctionDecl p f [Equation p lhs rhs']
 
-funLhs :: Parser Token (Ident, Lhs) a
+funLhs :: Parser a Token (Ident, Lhs)
 funLhs = mkFunLhs    <$> fun      <*> many1 pattern2
     <|?> flip ($ id) <$> pattern1 <*> opLhs
     <|?> curriedLhs
@@ -294,32 +290,32 @@ funLhs = mkFunLhs    <$> fun      <*> many1 pattern2
   mkOpLhs op t2 f t1      = (op, OpLhs (f t1) op t2)
   mkInfixPat op t2 f g t1 = f (g . InfixPattern t1 op) t2
 
-curriedLhs :: Parser Token (Ident,Lhs) a
+curriedLhs :: Parser a Token (Ident,Lhs)
 curriedLhs = apLhs <$> parens funLhs <*> many1 pattern2
   where apLhs (f, lhs) ts = (f, ApLhs lhs ts)
 
-declRhs :: Parser Token Rhs a
+declRhs :: Parser a Token Rhs
 declRhs = rhs equals
 
-rhs :: Parser Token a b -> Parser Token Rhs b
+rhs :: Parser a Token b -> Parser a Token Rhs
 rhs eq = rhsExpr <*> localDecls
   where rhsExpr =  SimpleRhs  <$-> eq <*> position <*> expr
                <|> GuardedRhs <$>  many1 (condExpr eq)
 
-localDecls :: Parser Token [Decl] a
+localDecls :: Parser a Token [Decl]
 localDecls = token KW_where <-*> layout valueDecls `opt` []
 
-valueDecls :: Parser Token [Decl] a
+valueDecls :: Parser a Token [Decl]
 valueDecls  = choice [infixDecl, valueDecl, foreignDecl] `sepBy` semicolon
 
-infixDecl :: Parser Token Decl a
+infixDecl :: Parser a Token Decl
 infixDecl = infixDeclLhs InfixDecl <*> option integer <*> funop `sepBy1` comma
 
-infixDeclLhs :: (Position -> Infix -> a) -> Parser Token a b
+infixDeclLhs :: (Position -> Infix -> a) -> Parser b Token a
 infixDeclLhs f = f <$> position <*> tokenOps infixKW
   where infixKW = [(KW_infix, Infix), (KW_infixl, InfixL), (KW_infixr, InfixR)]
 
-valueDecl :: Parser Token Decl a
+valueDecl :: Parser a Token Decl
 valueDecl = position <**> decl
   where
   decl =   var `sepBy1` comma         <**> valListDecl
@@ -343,7 +339,7 @@ valueDecl = position <**> decl
 
   isConstrId c = c == qConsId || isQualified c || isQTupleId c
 
-foreignDecl :: Parser Token Decl a
+foreignDecl :: Parser a Token Decl
 foreignDecl = ForeignDecl
           <$> tokenPos KW_foreign
           <*> callConv <*> (option string)
@@ -357,41 +353,41 @@ foreignDecl = ForeignDecl
 -- ---------------------------------------------------------------------------
 
 -- type0 ::= type1 ['->' type0]
-type0 :: Parser Token TypeExpr a
+type0 :: Parser a Token TypeExpr
 type0 = type1 `chainr1` (ArrowType <$-> token RightArrow)
 
 -- type1 ::= QTyCon { type2 } | type2
-type1 :: Parser Token TypeExpr a
+type1 :: Parser a Token TypeExpr
 type1 = ConstructorType <$> qtycon <*> many type2
      <|> type2 <\> qtycon
 
 -- type2 ::= anonType | identType | parenType | listType
-type2 :: Parser Token TypeExpr a
+type2 :: Parser a Token TypeExpr
 type2 = anonType <|> identType <|> parenType <|> listType
 
 -- anonType ::= '_'
-anonType :: Parser Token TypeExpr a
+anonType :: Parser a Token TypeExpr
 anonType = VariableType <$> anonIdent
 
 -- identType ::= <identifier>
-identType :: Parser Token TypeExpr a
+identType :: Parser a Token TypeExpr
 identType = VariableType <$> tyvar
         <|> flip ConstructorType [] <$> qtycon <\> tyvar
 
 -- parenType ::= '(' tupleType ')'
-parenType :: Parser Token TypeExpr a
+parenType :: Parser a Token TypeExpr
 parenType = parens tupleType
 
 -- tupleType ::= type0                         (parenthesized type)
 --            |  type0 ',' type0 { ',' type0 } (tuple type)
 --            |                                (unit type)
-tupleType :: Parser Token TypeExpr a
+tupleType :: Parser a Token TypeExpr
 tupleType = type0 <**> (tuple <$> many1 (comma <-*> type0) `opt` ParenType)
             `opt` TupleType []
   where tuple tys ty = TupleType (ty : tys)
 
 -- listType ::= '[' type0 ']'
-listType :: Parser Token TypeExpr a
+listType :: Parser a Token TypeExpr
 listType = ListType <$> brackets type0
 
 -- ---------------------------------------------------------------------------
@@ -402,7 +398,7 @@ listType = ListType <$> brackets type0
 --          |  <integer>
 --          |  <float>
 --          |  '"' <escaped string> '"'
-literal :: Parser Token Literal a
+literal :: Parser a Token Literal
 literal = mk Char   <$> char
       <|> mkInt     <$> integer
       <|> mk Float  <$> float
@@ -413,7 +409,7 @@ literal = mk Char   <$> char
 -- ---------------------------------------------------------------------------
 
 -- pattern0 ::= pattern1 [ gconop pattern0 ]
-pattern0 :: Parser Token Pattern a
+pattern0 :: Parser a Token Pattern
 pattern0 = pattern1 `chainr1` (flip InfixPattern <$> gconop)
 
 -- pattern1 ::= varId
@@ -422,7 +418,7 @@ pattern0 = pattern1 `chainr1` (flip InfixPattern <$> gconop)
 --           |  '-.' Float
 --           |  '(' parenPattern'
 --           | pattern2
-pattern1 :: Parser Token Pattern a
+pattern1 :: Parser a Token Pattern
 pattern1 =  varId <**> identPattern'            -- unqualified
         <|> qConId <\> varId <**> constrPattern -- qualified
         <|> minus     <**> negNum
@@ -449,26 +445,26 @@ pattern1 =  varId <**> identPattern'            -- unqualified
   gconPattern = ConstructorPattern <$> gconId <*-> rightParen
                                    <*> many pattern2
 
-pattern2 :: Parser Token Pattern a
+pattern2 :: Parser a Token Pattern
 pattern2 =  literalPattern <|> anonPattern <|> identPattern
         <|> parenPattern   <|> listPattern <|> lazyPattern
 
 -- literalPattern ::= <integer> | <char> | <float> | <string>
-literalPattern :: Parser Token Pattern a
+literalPattern :: Parser a Token Pattern
 literalPattern = LiteralPattern <$> literal
 
 -- anonPattern ::= '_'
-anonPattern :: Parser Token Pattern a
+anonPattern :: Parser a Token Pattern
 anonPattern = VariablePattern <$> anonIdent
 
 -- identPattern ::= Variable [ '@' pattern2 | '{' fields '}'
 --               |  qConId   [ '{' fields '}' ]
-identPattern :: Parser Token Pattern a
+identPattern :: Parser a Token Pattern
 identPattern =  varId <**> optAsRecPattern -- unqualified
             <|> qConId <\> varId <**> optRecPattern               -- qualified
 
 -- TODO: document me!
-parenPattern :: Parser Token Pattern a
+parenPattern :: Parser a Token Pattern
 parenPattern = leftParen <-*> parenPattern'
   where
   parenPattern' = minus  <**> minusPattern negNum
@@ -483,15 +479,15 @@ parenPattern = leftParen <-*> parenPattern'
 -- listPattern ::= '[' pattern0s ']'
 -- pattern0s   ::= {- empty -}
 --              |  pattern0 ',' pattern0s
-listPattern :: Parser Token Pattern a
+listPattern :: Parser a Token Pattern
 listPattern = mk' ListPattern <$> brackets (pattern0 `sepBy` comma)
 
 -- lazyPattern ::= '~' pattern2
-lazyPattern :: Parser Token Pattern a
+lazyPattern :: Parser a Token Pattern
 lazyPattern = mk LazyPattern <$-> token Tilde <*> pattern2
 
 -- optRecPattern ::= [ '{' fields '}' ]
-optRecPattern :: Parser Token (QualIdent -> Pattern) a
+optRecPattern :: Parser a Token (QualIdent -> Pattern)
 optRecPattern = mkRecPattern <$> fields pattern0 `opt` mkConPattern
   where
   mkRecPattern fs c = RecordPattern c fs
@@ -502,37 +498,37 @@ optRecPattern = mkRecPattern <$> fields pattern0 `opt` mkConPattern
 -- the left-hand side of a declaration.
 -- ---------------------------------------------------------------------------
 
-gconId :: Parser Token QualIdent a
+gconId :: Parser a Token QualIdent
 gconId = colon <|> tupleCommas
 
-negNum :: Parser Token (Ident -> Pattern) a
+negNum :: Parser a Token (Ident -> Pattern)
 negNum = flip NegativePattern <$> (mkInt <$> integer <|> mk Float <$> float)
 
-negFloat :: Parser Token (Ident -> Pattern) a
+negFloat :: Parser a Token (Ident -> Pattern)
 negFloat = flip NegativePattern . mk Float
            <$> (fromIntegral <$> integer <|> float)
 
-optAsRecPattern :: Parser Token (Ident -> Pattern) a
+optAsRecPattern :: Parser a Token (Ident -> Pattern)
 optAsRecPattern =  flip AsPattern <$-> token At <*> pattern2
                <|> recPattern     <$>  fields pattern0
                `opt` VariablePattern
   where recPattern fs v = RecordPattern (qualify v) fs
 
-optInfixPattern :: Parser Token (Pattern -> Pattern) a
+optInfixPattern :: Parser a Token (Pattern -> Pattern)
 optInfixPattern = mkInfixPat <$> gconop <*> pattern0
             `opt` id
   where mkInfixPat op t2 t1 = InfixPattern t1 op t2
 
-optTuplePattern :: Parser Token (Pattern -> Pattern) a
+optTuplePattern :: Parser a Token (Pattern -> Pattern)
 optTuplePattern = tuple <$> many1 (comma <-*> pattern0)
             `opt` ParenPattern
   where tuple ts t = mk TuplePattern (t:ts)
 
-parenMinusPattern :: Parser Token (Ident -> Pattern) a
-                  -> Parser Token (Ident -> Pattern) a
+parenMinusPattern :: Parser a Token (Ident -> Pattern)
+                  -> Parser a Token (Ident -> Pattern)
 parenMinusPattern p = p <.> optInfixPattern <.> optTuplePattern
 
-parenTuplePattern :: Parser Token Pattern a
+parenTuplePattern :: Parser a Token Pattern
 parenTuplePattern = pattern0 <**> optTuplePattern
               `opt` mk TuplePattern []
 
@@ -548,48 +544,48 @@ parenTuplePattern = pattern0 <**> optTuplePattern
 -- case a of { _ -> True :: Bool -> a }
 -- @
 -- can not be parsed with a limited parser lookahead.
-condExpr :: Parser Token a b -> Parser Token CondExpr b
+condExpr :: Parser a Token b -> Parser a Token CondExpr
 condExpr eq = CondExpr <$> position <*-> bar <*> expr0 <*-> eq <*> expr
 
 -- expr ::= expr0 [ '::' type0 ]
-expr :: Parser Token Expression a
+expr :: Parser a Token Expression
 expr = expr0 <??> (flip Typed <$-> token DoubleColon <*> type0)
 
 -- expr0 ::= expr1 { infixOp expr1 }
-expr0 :: Parser Token Expression a
+expr0 :: Parser a Token Expression
 expr0 = expr1 `chainr1` (flip InfixApply <$> infixOp)
 
 -- expr1 ::= - expr2 | -. expr2 | expr2
-expr1 :: Parser Token Expression a
+expr1 :: Parser a Token Expression
 expr1 =  UnaryMinus <$> (minus <|> fminus) <*> expr2
      <|> expr2
 
 -- expr2 ::= lambdaExpr | letExpr | doExpr | ifExpr | caseExpr | expr3
-expr2 :: Parser Token Expression a
+expr2 :: Parser a Token Expression
 expr2 = choice [ lambdaExpr, letExpr, doExpr, ifExpr, caseExpr
                , foldl1 Apply <$> many1 expr3
                ]
 
-expr3 :: Parser Token Expression a
+expr3 :: Parser a Token Expression
 expr3 = foldl RecordUpdate <$> expr4 <*> many recUpdate
   where recUpdate = layoutOff <-*> braces (field expr0 `sepBy1` comma)
 
-expr4 :: Parser Token Expression a
+expr4 :: Parser a Token Expression
 expr4 = choice
   [constant, anonFreeVariable, variable, parenExpr, listExpr]
 
-constant :: Parser Token Expression a
+constant :: Parser a Token Expression
 constant = Literal <$> literal
 
-anonFreeVariable :: Parser Token Expression a
+anonFreeVariable :: Parser a Token Expression
 anonFreeVariable =  (\ p v -> Variable $ qualify $ addPositionIdent p v)
                 <$> position <*> anonIdent
 
-variable :: Parser Token Expression a
+variable :: Parser a Token Expression
 variable = qFunId <**> optRecord
   where optRecord = flip Record <$> fields expr0 `opt` Variable
 
-parenExpr :: Parser Token Expression a
+parenExpr :: Parser a Token Expression
 parenExpr = parens pExpr
   where
   pExpr = (minus <|> fminus) <**> minusOrTuple
@@ -617,10 +613,10 @@ parenExpr = parens pExpr
   leftSection op f e = LeftSection (f e) op
   tuple es e = mk Tuple (e:es)
 
-infixOp :: Parser Token InfixOp a
+infixOp :: Parser a Token InfixOp
 infixOp = InfixOp <$> qfunop <|> InfixConstr <$> colon
 
-listExpr :: Parser Token Expression a
+listExpr :: Parser a Token Expression
 listExpr = brackets (elements `opt` mk' List [])
   where
   elements = expr <**> rest
@@ -636,36 +632,36 @@ listExpr = brackets (elements `opt` mk' List [])
   list es e2 e1 = mk' List (e1:e2:es)
   flip3 f x y z = f z y x
 
-lambdaExpr :: Parser Token Expression a
+lambdaExpr :: Parser a Token Expression
 lambdaExpr = mk Lambda <$-> token Backslash <*> many1 pattern2
                        <*-> expectRightArrow <*> expr
 
-letExpr :: Parser Token Expression a
+letExpr :: Parser a Token Expression
 letExpr = Let <$-> token KW_let <*> layout valueDecls
               <*-> (token KW_in <?> "in expected") <*> expr
 
-doExpr :: Parser Token Expression a
+doExpr :: Parser a Token Expression
 doExpr = uncurry Do <$-> token KW_do <*> layout stmts
 
-ifExpr :: Parser Token Expression a
+ifExpr :: Parser a Token Expression
 ifExpr = mk IfThenElse
     <$-> token KW_if                         <*> expr
     <*-> (token KW_then <?> "then expected") <*> expr
     <*-> (token KW_else <?> "else expected") <*> expr
 
-caseExpr :: Parser Token Expression a
+caseExpr :: Parser a Token Expression
 caseExpr = keyword <*> expr
       <*-> (token KW_of <?> "of expected") <*> layout (alt `sepBy1` semicolon)
   where keyword =  mk Case Flex  <$-> token KW_fcase
                <|> mk Case Rigid <$-> token KW_case
 
-alt :: Parser Token Alt a
+alt :: Parser a Token Alt
 alt = Alt <$> position <*> pattern0 <*> rhs expectRightArrow
 
-fields :: Parser Token a b -> Parser Token [Field a] b
+fields :: Parser a Token b -> Parser a Token [Field b]
 fields p = layoutOff <-*> braces (field p `sepBy` comma)
 
-field :: Parser Token a b -> Parser Token (Field a) b
+field :: Parser a Token b -> Parser a Token (Field b)
 field p = Field <$> position <*> qfun <*-> expectEquals <*> p
 
 -- ---------------------------------------------------------------------------
@@ -679,32 +675,32 @@ field p = Field <$> position <*> qfun <*-> expectEquals <*> p
 -- prefix of a let expression.
 -- ---------------------------------------------------------------------------
 
-stmts :: Parser Token ([Statement], Expression) a
+stmts :: Parser a Token ([Statement], Expression)
 stmts = stmt reqStmts optStmts
 
-reqStmts :: Parser Token (Statement -> ([Statement], Expression)) a
+reqStmts :: Parser a Token (Statement -> ([Statement], Expression))
 reqStmts = (\ (sts, e) st -> (st : sts, e)) <$-> semicolon <*> stmts
 
-optStmts :: Parser Token (Expression -> ([Statement],Expression)) a
+optStmts :: Parser a Token (Expression -> ([Statement],Expression))
 optStmts = succeed (mk StmtExpr) <.> reqStmts `opt` (,) []
 
-quals :: Parser Token [Statement] a
+quals :: Parser a Token [Statement]
 quals = stmt (succeed id) (succeed $ mk StmtExpr) `sepBy1` comma
 
-stmt :: Parser Token (Statement -> a) b
-     -> Parser Token (Expression -> a) b -> Parser Token a b
+stmt :: Parser a Token (Statement -> b)
+     -> Parser a Token (Expression -> b) -> Parser a Token b
 stmt stmtCont exprCont =  letStmt stmtCont exprCont
                       <|> exprOrBindStmt stmtCont exprCont
 
-letStmt :: Parser Token (Statement -> a) b
-        -> Parser Token (Expression -> a) b -> Parser Token a b
+letStmt :: Parser a Token (Statement -> b)
+        -> Parser a Token (Expression -> b) -> Parser a Token b
 letStmt stmtCont exprCont = token KW_let <-*> layout valueDecls <**> optExpr
   where optExpr =  flip Let <$-> token KW_in <*> expr <.> exprCont
                <|> succeed StmtDecl <.> stmtCont
 
-exprOrBindStmt :: Parser Token (Statement -> a) b
-               -> Parser Token (Expression -> a) b
-               -> Parser Token a b
+exprOrBindStmt :: Parser a Token (Statement -> b)
+               -> Parser a Token (Expression -> b)
+               -> Parser a Token b
 exprOrBindStmt stmtCont exprCont =
        mk StmtBind <$> pattern0 <*-> leftArrow <*> expr <**> stmtCont
   <|?> expr <\> token KW_let <**> exprCont
@@ -713,137 +709,137 @@ exprOrBindStmt stmtCont exprCont =
 -- Goals
 -- ---------------------------------------------------------------------------
 
-goal :: Parser Token Goal a
+goal :: Parser a Token Goal
 goal = Goal <$> position <*> expr <*> localDecls
 
 -- ---------------------------------------------------------------------------
 -- Literals, identifiers, and (infix) operators
 -- ---------------------------------------------------------------------------
 
-char :: Parser Token Char a
+char :: Parser a Token Char
 char = cval <$> token CharTok
 
-float :: Parser Token Double a
+float :: Parser a Token Double
 float = fval <$> token FloatTok
 
-int :: Parser Token Int a
+int :: Parser a Token Int
 int = fromInteger <$> integer
 
-integer :: Parser Token Integer a
+integer :: Parser a Token Integer
 integer = ival <$> token IntTok
 
-string :: Parser Token String a
+string :: Parser a Token String
 string = sval <$> token StringTok
 
-tycon :: Parser Token Ident a
+tycon :: Parser a Token Ident
 tycon = conId
 
-anonOrTyvar :: Parser Token Ident a
+anonOrTyvar :: Parser a Token Ident
 anonOrTyvar = anonIdent <|> tyvar
 
-tyvar :: Parser Token Ident a
+tyvar :: Parser a Token Ident
 tyvar = varId
 
-qtycon :: Parser Token QualIdent a
+qtycon :: Parser a Token QualIdent
 qtycon = qConId
 
-varId :: Parser Token Ident a
+varId :: Parser a Token Ident
 varId = ident
 
-funId :: Parser Token Ident a
+funId :: Parser a Token Ident
 funId = ident
 
-conId :: Parser Token Ident a
+conId :: Parser a Token Ident
 conId = ident
 
-funSym :: Parser Token Ident a
+funSym :: Parser a Token Ident
 funSym = sym
 
-conSym :: Parser Token Ident a
+conSym :: Parser a Token Ident
 conSym = sym
 
-modIdent :: Parser Token ModuleIdent a
+modIdent :: Parser a Token ModuleIdent
 modIdent = mIdent <?> "module name expected"
 
-var :: Parser Token Ident a
+var :: Parser a Token Ident
 var = varId <|> parens (funSym <?> "operator symbol expected")
 
-fun :: Parser Token Ident a
+fun :: Parser a Token Ident
 fun = funId <|> parens (funSym <?> "operator symbol expected")
 
-con :: Parser Token Ident a
+con :: Parser a Token Ident
 con = conId <|> parens (conSym <?> "operator symbol expected")
 
-funop :: Parser Token Ident a
+funop :: Parser a Token Ident
 funop = funSym <|> backquotes (funId <?> "operator name expected")
 
-conop :: Parser Token Ident a
+conop :: Parser a Token Ident
 conop = conSym <|> backquotes (conId <?> "operator name expected")
 
-qFunId :: Parser Token QualIdent a
+qFunId :: Parser a Token QualIdent
 qFunId = qIdent
 
-qConId :: Parser Token QualIdent a
+qConId :: Parser a Token QualIdent
 qConId = qIdent
 
-qFunSym :: Parser Token QualIdent a
+qFunSym :: Parser a Token QualIdent
 qFunSym = qSym
 
-qConSym :: Parser Token QualIdent a
+qConSym :: Parser a Token QualIdent
 qConSym = qSym
 
-gConSym :: Parser Token QualIdent a
+gConSym :: Parser a Token QualIdent
 gConSym = qConSym <|> colon
 
-qfun :: Parser Token QualIdent a
+qfun :: Parser a Token QualIdent
 qfun = qFunId <|> parens (qFunSym <?> "operator symbol expected")
 
-qfunop :: Parser Token QualIdent a
+qfunop :: Parser a Token QualIdent
 qfunop = qFunSym <|> backquotes (qFunId <?> "operator name expected")
 
-gconop :: Parser Token QualIdent a
+gconop :: Parser a Token QualIdent
 gconop = gConSym <|> backquotes (qConId <?> "operator name expected")
 
-anonIdent :: Parser Token Ident a
+anonIdent :: Parser a Token Ident
 anonIdent = (\ p -> addPositionIdent p anonId) <$> tokenPos Underscore
 
-mIdent :: Parser Token ModuleIdent a
+mIdent :: Parser a Token ModuleIdent
 mIdent = mIdent' <$> position <*>
      tokens [Id,QId,Id_as,Id_ccall,Id_forall,Id_hiding,
              Id_interface,Id_primitive,Id_qualified]
   where mIdent' p a = addPositionModuleIdent p $
                       mkMIdent (modulVal a ++ [sval a])
 
-ident :: Parser Token Ident a
+ident :: Parser a Token Ident
 ident = (\ pos -> mkIdentPosition pos . sval) <$> position <*>
        tokens [Id,Id_as,Id_ccall,Id_forall,Id_hiding,
                Id_interface,Id_primitive,Id_qualified]
 
-qIdent :: Parser Token QualIdent a
+qIdent :: Parser a Token QualIdent
 qIdent =  qualify  <$> ident
       <|> mkQIdent <$> position <*> token QId
   where mkQIdent p a = qualifyWith (mkMIdent (modulVal a))
                                    (mkIdentPosition p (sval a))
 
-sym :: Parser Token Ident a
+sym :: Parser a Token Ident
 sym = (\ pos -> mkIdentPosition pos . sval) <$> position <*>
       tokens [Sym, SymDot, SymMinus, SymMinusDot]
 
-qSym :: Parser Token QualIdent a
+qSym :: Parser a Token QualIdent
 qSym = qualify <$> sym <|> mkQIdent <$> position <*> token QSym
   where mkQIdent p a = qualifyWith (mkMIdent (modulVal a))
                                    (mkIdentPosition p (sval a))
 
-colon :: Parser Token QualIdent a
+colon :: Parser a Token QualIdent
 colon = (\ p -> qualify $ addPositionIdent p consId) <$> tokenPos Colon
 
-minus :: Parser Token Ident a
+minus :: Parser a Token Ident
 minus = (\ p -> addPositionIdent p minusId) <$> tokenPos SymMinus
 
-fminus :: Parser Token Ident a
+fminus :: Parser a Token Ident
 fminus = (\ p -> addPositionIdent p fminusId) <$> tokenPos SymMinusDot
 
-tupleCommas :: Parser Token QualIdent a
+tupleCommas :: Parser a Token QualIdent
 tupleCommas = (\ p -> qualify . addPositionIdent p . tupleId . succ . length)
               <$> position <*> many1 comma
 
@@ -853,96 +849,93 @@ tupleCommas = (\ p -> qualify . addPositionIdent p . tupleId . succ . length)
 
 -- |This function starts a new layout block but does not wait for its end.
 -- This is only used for parsing the module header.
-startLayout :: Parser Token a b -> Parser Token a b
-startLayout p = layoutOff <-*> leftBraceSemicolon <-*> p
+startLayout :: Parser a Token b -> Parser a Token b
+startLayout p = layoutOff <-*> leftBrace <-*> p
              <|> layoutOn <-*> p
 
-layout :: Parser Token a b -> Parser Token a b
-layout p =  layoutOff <-*> between leftBraceSemicolon p rightBrace
+layout :: Parser a Token b -> Parser a Token b
+layout p =  layoutOff <-*> braces p
         <|> layoutOn  <-*> p <*-> (token VRightBrace <|> layoutEnd)
 
 -- ---------------------------------------------------------------------------
 -- Bracket combinators
 -- ---------------------------------------------------------------------------
 
-braces :: Parser Token a b -> Parser Token a b
+braces :: Parser a Token b -> Parser a Token b
 braces p = between leftBrace p rightBrace
 
-brackets :: Parser Token a b -> Parser Token a b
+brackets :: Parser a Token b -> Parser a Token b
 brackets p = between leftBracket p rightBracket
 
-parens :: Parser Token a b -> Parser Token a b
+parens :: Parser a Token b -> Parser a Token b
 parens p = between leftParen p rightParen
 
-backquotes :: Parser Token a b -> Parser Token a b
+backquotes :: Parser a Token b -> Parser a Token b
 backquotes p = between backquote p expectBackquote
 
 -- ---------------------------------------------------------------------------
 -- Simple token parsers
 -- ---------------------------------------------------------------------------
 
-token :: Category -> Parser Token Attributes a
+token :: Category -> Parser a Token Attributes
 token c = attr <$> symbol (Token c NoAttributes)
   where attr (Token _ a) = a
 
-tokens :: [Category] -> Parser Token Attributes a
+tokens :: [Category] -> Parser a Token Attributes
 tokens = foldr1 (<|>) . map token
 
-tokenPos :: Category -> Parser Token Position a
+tokenPos :: Category -> Parser a Token Position
 tokenPos c = position <*-> token c
 
-tokenOps :: [(Category, a)] -> Parser Token a b
+tokenOps :: [(Category, b)] -> Parser a Token b
 tokenOps cs = ops [(Token c NoAttributes, x) | (c, x) <- cs]
 
-comma :: Parser Token Attributes a
+comma :: Parser a Token Attributes
 comma = token Comma
 
-semicolon :: Parser Token Attributes a
+semicolon :: Parser a Token Attributes
 semicolon = token Semicolon <|> token VSemicolon
 
-bar :: Parser Token Attributes a
+bar :: Parser a Token Attributes
 bar = token Bar
 
-equals :: Parser Token Attributes a
+equals :: Parser a Token Attributes
 equals = token Equals
 
-expectEquals :: Parser Token Attributes a
+expectEquals :: Parser a Token Attributes
 expectEquals = equals <?> "= expected"
 
-expectWhere :: Parser Token Attributes a
+expectWhere :: Parser a Token Attributes
 expectWhere = token KW_where <?> "where expected"
 
-expectRightArrow :: Parser Token Attributes a
+expectRightArrow :: Parser a Token Attributes
 expectRightArrow  = token RightArrow <?> "-> expected"
 
-backquote :: Parser Token Attributes a
+backquote :: Parser a Token Attributes
 backquote = token Backquote
 
-expectBackquote :: Parser Token Attributes a
+expectBackquote :: Parser a Token Attributes
 expectBackquote = backquote <?> "backquote (`) expected"
 
-leftParen :: Parser Token Attributes a
+leftParen :: Parser a Token Attributes
 leftParen = token LeftParen
 
-rightParen :: Parser Token Attributes a
+rightParen :: Parser a Token Attributes
 rightParen = token RightParen
 
-leftBracket :: Parser Token Attributes a
+leftBracket :: Parser a Token Attributes
 leftBracket = token LeftBracket
 
-rightBracket :: Parser Token Attributes a
+rightBracket :: Parser a Token Attributes
 rightBracket = token RightBracket
 
-leftBrace :: Parser Token Attributes a
+leftBrace :: Parser a Token Attributes
 leftBrace = token LeftBrace
 
-leftBraceSemicolon :: Parser Token Attributes a
-leftBraceSemicolon = token LeftBraceSemicolon
-
-rightBrace :: Parser Token Attributes a
+rightBrace :: Parser a Token Attributes
 rightBrace = token RightBrace
 
-leftArrow :: Parser Token Attributes a
+leftArrow :: Parser a Token Attributes
 leftArrow = token LeftArrow
 
 -- ---------------------------------------------------------------------------